requirements.txt
python-dotenv==1.0.0
Flask==2.3.2
Flask_Cors==3.0.10
google_api_python_client==2.86.0
langchain==0.0.180
PyDrive==1.3.1
PyPDF2==3.0.1
faiss-cpu
openai
tiktoken
waitress==2.1.2
pickle5 